Refactor config module: move defaults to separate file
This commit restructures the config module for better organization:

Changes:
- Move src/config.rs → src/config/mod.rs
- Extract defaults module → src/config/defaults.rs
- Update mod.rs to import defaults as a submodule

Benefits:
- Cleaner separation of concerns
- Easier to locate and modify default values
- Better file organization as config module grows
- No functional changes, pure refactor

File structure:
src/config/
├── mod.rs       - Main config module
└── defaults.rs  - Default configuration values
